from home

+++ functions.py
@@ -29,14 +29,19 @@ except:
       
 
 
-def insertValue(intData, intTime = datetime.date.today()):
+def insertValue(intData, intTime = datetime.date.today()):    
     intData = intData.lower().strip()
-    with open(FILENAME, "a", encoding = "utf8") as file:
+    searchFile = searchInt(intData)
+    if searchFile == False:
+        with open(FILENAME, "a", encoding = "utf8") as file:
             writeNames = csv.writer(file)
             writeNames.writerow([intData, intTime])
+    else:
+        print("Указанное имя существует")
     
 def searchInt(name):
     name = name.lower().strip()
+    elem = False
     with open(FILENAME, "r") as file:
         freader = csv.reader(file)
         for row in freader:
@@ -46,18 +51,30 @@ def searchInt(name):
 
 def deleteInt(name):
     elem = searchInt(name)
-    with open("docs/products2.csv", 'w', newline = "") as newFile:
-        writer = csv.writer(newFile)
-        with open(FILENAME, "r", newline = "") as file:
-            freader = csv.reader(file)
-            for row in freader:
-                model = [row[0], row[1]]
-                if model == elem:
-                    continue
-                writer.writerow(model)
-    os.remove("docs/products.csv")
-    os.rename("docs/products2.csv", "docs/products.csv")
+    if elem != False:
+        with open("docs/products2.csv", 'w', newline = "") as newFile:
+            writer = csv.writer(newFile)
+            with open(FILENAME, "r", newline = "") as file:
+                freader = csv.reader(file)
+                for row in freader:
+                    model = [row[0], row[1]]
+                    if model == elem:
+                        continue
+                    writer.writerow(model)
+        os.remove("docs/products.csv")
+        os.rename("docs/products2.csv", "docs/products.csv")
+    else:
+        print("Такого имени нет")
 
+def changeInt(intData, intTime):
+    intData = intData.lower().strip()
+    elem = searchInt(intData)
+    if elem != False and intTime != "":
+        deleteInt(intData)
+        insertValue(intData, intTime)
+    else:
+        print("Такого имени нет")
+
